ENH: Create a more complete Slits class
